// src/app/store/actionTypes/tag-action.types.ts
export enum TagActionTypes {
    CREATE_TAG = '[Tags] create tag',  
    CREATE_TAG_OK = '[Tags] create tag ok',  
    CREATE_TAG_FAIL = '[Tags] create tag fail',

    GET_TAGS = '[Tags] get tags',
    GET_TAGS_OK = '[Tags] get tags ok',
    GET_TAGS_FAIL = '[Tags] get tags fail',

    CLEAR_CURRENT_TAG = '[Tags] clear tag',

    DELETE_TAG = '[Tags] delete tag',
    DELETE_TAG_OK = '[Tags] delete tag ok',
    DELETE_TAG_FAIL = '[Tags] delete tag fail',

    GET_TAG = '[Tags] get tag',
    GET_TAG_OK = '[Tags] get tag ok',
    GET_TAG_FAIL = '[Tags] get tag fail',

    UPDATE_TAG = '[Tags] update tag',
    UPDATE_TAG_OK = '[Tags] update tag ok',
    UPDATE_TAG_FAIL = '[Tags] update tag fail',
}